Разбил скрипт на модули
