+++ js/map.js
@@ -1,0 +1,1 @@
+//Здесь будет карта
